fix: same title entities
