+++ admin/src/pages/HomePage/components/NavigationItemForm/index.tsx
@@ -1,39 +1,16 @@
-import { AnyEntity, Field } from '@sensinum/strapi-utils';
+import { AnyEntity } from '@sensinum/strapi-utils';
 import { Form, useNotification } from '@strapi/strapi/admin';
-import { get, isBoolean, isEmpty, isNil, isObject, isString, set, sortBy } from 'lodash';
-import { useCallback, useEffect, useMemo, useState } from 'react';
+import { useCallback, useEffect, useState } from 'react';
 import { useIntl } from 'react-intl';
 
-import { Grid, Modal, Toggle } from '@strapi/design-system';
+import { Grid, Modal } from '@strapi/design-system';
 
-import {
-  Box,
-  Button,
-  Divider,
-  MultiSelect,
-  MultiSelectOption,
-  SingleSelect,
-  SingleSelectOption,
-  TextInput,
-} from '@strapi/design-system';
 import { NavigationSchema } from '../../../../api/validators';
 import { NavigationItemAdditionalField } from '../../../../schemas';
 import { getTrad } from '../../../../translations';
-import {
-  Effect,
-  FormChangeEvent,
-  FormItemErrorSchema,
-  ToBeFixed,
-  VoidEffect,
-} from '../../../../types';
-import {
-  useConfig,
-  useContentTypeItems,
-  useCopyNavigationItemI18n,
-  useNavigations,
-} from '../../hooks';
+import { Effect, FormItemErrorSchema, VoidEffect } from '../../../../types';
+import { useConfig, useContentTypeItems } from '../../hooks';
 import { extractRelatedItemLabel } from '../../utils';
-import { AdditionalFieldInput } from '../AdditionalFieldInput';
 import { NavigationItemPopupFooter } from '../NavigationItemPopup/NavigationItemPopupFooter';
 import { ContentTypeEntity } from './types';
 import {
@@ -41,9 +18,18 @@ import {
   navigationItemFormSchema,
   NavigationItemFormSchema,
 } from './utils/form';
-import { useSlug } from './utils/hooks';
-import { generatePreviewPath, generateUiRouterKey } from './utils/properties';
-import { ControllableCombobox } from '../ControllableCombobox';
+import { useFormValues, usePayload, useSlug } from './utils/hooks';
+import { generateUiRouterKey } from './utils/properties';
+import { NavigationItemFormContext } from './context/NavigationItemFormContext';
+import { NavigationItemTypeField } from './components/NavigationItemTypeField';
+import { AttachToMenuField } from './components/AttachToMenuField';
+import { TitleField } from './components/TitleField';
+import { ReadFieldsFromRelatedField } from './components/ReadFieldsFromRelatedField';
+import { PathField } from './components/PathField';
+import { AdditionalFields } from './components/AdditionalFields';
+import { CopyFromLocales } from './components/CopyFromLocales';
+import { RelatedTypeField } from './components/RelatedTypeField';
+import { RelatedEntityField } from './components/RelatedEntityField';
 
 export { ContentTypeEntity, GetContentTypeEntitiesPayload } from './types';
 export { NavigationItemFormSchema } from './utils/form';
@@ -83,132 +69,33 @@ export const NavigationItemForm: React.FC<NavigationItemFormProps> = ({
 
   const [isSingleSelected, setIsSingleSelected] = useState(false);
 
-  const [itemLocaleCopyValue, setItemLocaleCopyValue] = useState<string>();
-
-  const [formValue, setFormValue] = useState<NavigationItemFormSchema>(
-    {} as NavigationItemFormSchema
-  );
-
-  const [formError, setFormError] = useState<FormItemErrorSchema<NavigationItemFormSchema>>();
-
   const configQuery = useConfig();
-
-  const availableAudiences = configQuery.data?.availableAudience ?? [];
 
   const contentTypes = configQuery.data?.contentTypes ?? [];
 
   const { toggleNotification } = useNotification();
 
-  const copyItemFromLocaleMutation = useCopyNavigationItemI18n();
+  const {
+    formValue,
+    renderError,
+    setFormError,
+    handleChange,
+    setFormValue,
+    setFormValueItem,
+    setFormValuesItems,
+  } = useFormValues();
 
-  const navigationsQuery = useNavigations();
+  const { encodePayload, decodePayload } = usePayload();
 
-  const handleChange = (
-    eventOrPath: FormChangeEvent,
-    value?: any,
-    nativeOnChange?: (eventOrPath: FormChangeEvent, value?: any) => void
-  ) => {
-    if (nativeOnChange) {
-      let fieldName = eventOrPath;
-      let fieldValue = value;
-      if (isObject(eventOrPath)) {
-        const { name: targetName, value: targetValue } = eventOrPath.target;
-        fieldName = targetName;
-        fieldValue = isNil(fieldValue) ? targetValue : fieldValue;
-      }
-      if (isString(fieldName)) {
-        setFormValueItem(fieldName, fieldValue);
-      }
+  const isExternal = formValue.type === 'EXTERNAL';
+  const isInternal = formValue.type === 'INTERNAL';
 
-      return nativeOnChange(eventOrPath, fieldValue);
-    }
-  };
+  const submitDisabled = (isInternal && !formValue.related && !isSingleSelected) || isLoading;
 
-  const setFormValueItem = (path: string, value: any) => {
-    setFormValue(
-      set(
-        {
-          ...formValue,
-          additionalFields: {
-            ...formValue.additionalFields,
-          },
-          updated: true,
-        },
-        path,
-        value
-      )
-    );
-  };
-
-  const setFormValuesItems = (values: any) =>
-    setFormValue({
-      ...formValue,
-      additionalFields: {
-        ...formValue.additionalFields,
-      },
-      updated: true,
-      ...values,
-    });
-
-  const encodePayload = (values: NavigationItemFormSchema): NavigationItemFormSchema => {
-    return {
-      ...values,
-      additionalFields:
-        configQuery.data?.additionalFields.reduce((acc, field: NavigationItemAdditionalField) => {
-          const { name, type } = field as ToBeFixed;
-
-          if (name in (values.additionalFields ?? {})) {
-            let val = values.additionalFields[name];
-
-            switch (type) {
-              case 'boolean':
-                val = isBoolean(val) ? `${val}` : val;
-                break;
-              case 'media':
-                val = val ? JSON.stringify(val) : val;
-                break;
-              default:
-                break;
-            }
-            return {
-              ...acc,
-              [name]: val,
-            };
-          }
-          return acc;
-        }, {}) || {},
-    };
-  };
-
-  const decodePayload = (values: NavigationItemFormSchema): NavigationItemFormSchema => {
-    return {
-      ...values,
-      additionalFields:
-        configQuery.data?.additionalFields.reduce((acc, field: NavigationItemAdditionalField) => {
-          const { name, type } = field as ToBeFixed;
-
-          if (name in (values.additionalFields ?? {})) {
-            let val = values.additionalFields[name];
-
-            switch (type) {
-              case 'boolean':
-                val = val === 'true' ? true : false;
-                break;
-              case 'media':
-                val = val ? JSON.parse(val) : val;
-                break;
-              default:
-                break;
-            }
-            return {
-              ...acc,
-              [name]: val,
-            };
-          }
-          return acc;
-        }, {}) || {},
-    };
-  };
+  const contentTypeItemsQuery = useContentTypeItems({
+    uid: isInternal ? formValue.relatedType : '',
+    locale,
+  });
 
   const submit = async (e: React.MouseEvent, values: NavigationItemFormSchema) => {
     e.preventDefault();
@@ -280,58 +167,7 @@ export const NavigationItemForm: React.FC<NavigationItemFormProps> = ({
     }
   };
 
-  const renderError = (field: string, messageKey?: string): string | undefined => {
-    return get(formError, field) ? formatMessage(getTrad(messageKey ?? field)) : undefined;
-  };
-
-  const initialRelatedTypeSelected = current.type === 'INTERNAL' ? current.relatedType : undefined;
-  const {
-    path: currentPath,
-    type: currentType,
-    title: currentTitle,
-    autoSync: autoSyncEnabled,
-  } = formValue;
-
-  const { relatedType: currentRelatedType, related: currentRelated } =
-    formValue.type === 'INTERNAL'
-      ? formValue
-      : {
-          related: undefined,
-          relatedType: undefined,
-        };
-
-  const isExternal = currentType === 'EXTERNAL';
-  const isInternal = currentType === 'INTERNAL';
-
-  const pathSourceName = isExternal ? 'externalPath' : 'path';
-
-  const submitDisabled = (isInternal && !isSingleSelected && isNil(currentRelated)) || isLoading;
-
-  const contentTypeItemsQuery = useContentTypeItems({
-    uid: currentRelatedType ?? '',
-    locale,
-  });
-
   const slugifyMutation = useSlug();
-
-  const availableLocaleOptions = useMemo(
-    () =>
-      availableLocale.map((locale, index) => ({
-        key: `${locale}-${index}`,
-        value: locale,
-        label: locale,
-      })),
-    [availableLocale]
-  );
-
-  const audienceOptions = useMemo(
-    () =>
-      availableAudiences.map((item) => ({
-        value: item.documentId ?? 0,
-        label: item.name ?? ' ',
-      })),
-    [availableAudiences]
-  );
 
   const getDefaultTitle = useCallback(
     (related: string | undefined, relatedType: string | undefined, isSingleSelected: boolean) => {
@@ -360,112 +196,6 @@ export const NavigationItemForm: React.FC<NavigationItemFormProps> = ({
     [contentTypeItemsQuery.data, configQuery.data, contentTypes]
   );
 
-  const availableNavigationItemTypeOptions = !configQuery.data?.contentTypes.length
-    ? (['EXTERNAL', 'WRAPPER'] as const)
-    : (['INTERNAL', 'EXTERNAL', 'WRAPPER'] as const);
-  const navigationItemTypeOptions = availableNavigationItemTypeOptions.map((key) => {
-    return {
-      key,
-      value: key,
-      label: formatMessage(getTrad(`popup.item.form.type.${key.toLowerCase()}.label`)),
-    };
-  });
-
-  const relatedSelectOptions = sortBy(
-    contentTypeItemsQuery.data?.map((item) => {
-      const label = appendLabelPublicationStatus(
-        extractRelatedItemLabel(
-          {
-            ...item,
-            __collectionUid: currentRelatedType,
-          },
-          configQuery.data
-        ),
-        item,
-        false
-      );
-
-      return {
-        key: item?.documentId?.toString(),
-        value: item?.documentId?.toString(),
-        label: label,
-      };
-    }) ?? [],
-    (item) => item.label
-  );
-
-  const relatedTypeSelectOptions = useMemo(
-    () =>
-      sortBy(
-        configQuery.data?.contentTypes
-          ?.filter((contentType) => {
-            if (contentType.isSingle) {
-              return !!(
-                currentRelatedType &&
-                [currentRelatedType, initialRelatedTypeSelected].includes(contentType.uid)
-              );
-            }
-
-            return true;
-          })
-          .map((item) => {
-            return {
-              key: item.uid,
-              value: item.uid,
-              label: item.contentTypeName,
-            };
-          }),
-        (item) => item.label
-      ),
-    [configQuery.data, currentRelatedType]
-  );
-
-  const thereAreNoMoreContentTypes = isEmpty(relatedSelectOptions);
-
-  const onCopyFromLocale = useCallback(
-    async (event: React.BaseSyntheticEvent) => {
-      event.preventDefault();
-      event.stopPropagation();
-
-      const source = navigationsQuery.data?.find(({ locale }) => locale === itemLocaleCopyValue);
-
-      if (source) {
-        setIsLoading(true);
-
-        copyItemFromLocaleMutation.mutate(
-          {
-            target: currentNavigation.documentId,
-            structureId: current.structureId,
-            source: source.documentId,
-          },
-          {
-            onSuccess(data) {
-              copyItemFromLocaleMutation.reset();
-
-              const { type, externalPath, path, related, title, uiRouterKey } = data;
-              const { __type, documentId } = related ?? {};
-
-              setFormValueItem('type', type);
-              setFormValueItem('externalPath', externalPath ?? undefined);
-              setFormValueItem('path', path ?? undefined);
-              setFormValueItem('title', title);
-              setFormValueItem('uiRouterKey', uiRouterKey);
-
-              if (__type && documentId) {
-                setFormValueItem('related', documentId);
-                setFormValueItem('relatedType', __type);
-              }
-            },
-            onSettled() {
-              setIsLoading(false);
-            },
-          }
-        );
-      }
-    },
-    [setIsLoading, copyItemFromLocaleMutation, navigationsQuery]
-  );
-
   useEffect(() => {
     setFormValue(
       decodePayload({
@@ -475,490 +205,75 @@ export const NavigationItemForm: React.FC<NavigationItemFormProps> = ({
     );
   }, [current]);
 
-  useEffect(() => {
-    if (currentRelatedType) {
-      const relatedType = configQuery.data?.contentTypes.find(
-        (contentType) => contentType.uid === currentRelatedType
-      );
-
-      if (relatedType) {
-        setIsSingleSelected(relatedType.isSingle);
-
-        if (relatedType.isSingle && contentTypeItemsQuery.data?.length) {
-          const nextRelated = contentTypeItemsQuery.data[0];
-
-          if (nextRelated) {
-            setFormValueItem('related', nextRelated.documentId);
-          }
-        }
-      }
-    }
-  }, [currentRelatedType, configQuery.data, contentTypeItemsQuery.data]);
-
-  useEffect(() => {
-    if (
-      autoSyncEnabled &&
-      currentType === 'INTERNAL' &&
-      currentRelated &&
-      currentRelatedType &&
-      configQuery.data
-    ) {
-      const relatedItem = contentTypeItemsQuery.data?.find((item) => {
-        return item.documentId === currentRelated;
-      });
-
-      if (relatedItem) {
-        const { contentTypesNameFields, pathDefaultFields } = configQuery.data;
-
-        const nextPath = pathDefaultFields[currentRelatedType]?.reduce<string | undefined>(
-          (acc, field) => {
-            return acc ? acc : relatedItem?.[field];
-          },
-          undefined
-        );
-
-        const nextTitle = (contentTypesNameFields[currentRelatedType] ?? [])
-          .concat(contentTypesNameFields.default ?? [])
-          .reduce<undefined | string>((acc, field) => {
-            return acc ? acc : relatedItem?.[field];
-          }, undefined);
-
-        const batch: Array<{ name: keyof NavigationItemFormSchema; value: string }> = [];
-
-        if (nextPath && nextPath !== currentPath) {
-          batch.push({ name: 'path', value: nextPath });
-        }
-
-        if (nextTitle && nextTitle !== currentTitle) {
-          batch.push({ name: 'title', value: nextTitle });
-        }
-
-        setTimeout(() => {
-          batch.forEach((next) => {
-            setFormValueItem(next.name as any, next.value);
-          });
-        }, 100);
-      }
-    }
-  }, [
-    currentTitle,
-    currentPath,
-    autoSyncEnabled,
-    currentType,
-    currentRelated,
-    currentRelatedType,
-    configQuery.data,
-    contentTypeItemsQuery.data,
-  ]);
-
-  useEffect(() => {
-    if (currentType === 'INTERNAL' && currentRelatedType === '') {
-      setFormValueItem("relatedType", configQuery.data?.defaultContentType);
-    }
-  }, [configQuery.data?.defaultContentType, currentType, currentRelatedType]);
-
   return (
     <>
       <Modal.Body>
         <Form width="auto" height="auto" method="POST" initialValues={formValue}>
           {({ values, onChange }) => {
-            const internalValues =
-              values.type === 'INTERNAL'
-                ? values
-                : {
-                    related: undefined,
-                    relatedType: undefined,
-                  };
-
-            const pathDefault = generatePreviewPath({
-              currentPath: values.path,
-              isExternal: values.type === 'EXTERNAL',
-              current,
-              currentType: values.type,
-              config: configQuery.data,
-              contentTypeItems: contentTypeItemsQuery.data,
-              currentRelated: internalValues.related,
-              currentRelatedType: internalValues.relatedType,
-              isSingleSelected,
-            });
-
             return (
-              <>
+              <NavigationItemFormContext.Provider
+                value={{
+                  values,
+                  onChange,
+                  handleChange,
+                  renderError,
+                  setFormValueItem,
+                  canUpdate,
+                  isLoading,
+                }}
+              >
                 <Grid.Root gap={5} paddingBottom={1}>
-                  <Grid.Item alignItems="flex-start" key="title" col={8}>
-                    <Field
-                      name="type"
-                      label={formatMessage(getTrad('popup.item.form.type.label', 'Internal link'))}
-                      error={renderError('type')}
-                      hint={formatMessage(
-                        getTrad('popup.item.form.title.placeholder', 'e.g. Blog')
-                      )}
-                    >
-                      <SingleSelect
-                        onChange={(eventOrPath: FormChangeEvent) =>
-                          handleChange('type', eventOrPath, onChange)
-                        }
-                        value={values.type}
-                        name="type"
-                        disabled={!canUpdate}
-                        width="100%"
-                      >
-                        {navigationItemTypeOptions.map(({ key, label, value }) => (
-                          <SingleSelectOption key={key} value={value}>
-                            {label}
-                          </SingleSelectOption>
-                        ))}
-                      </SingleSelect>
-                    </Field>
-                  </Grid.Item>
-
-                  <Grid.Item alignItems="flex-start" key="menuAttached" col={4}>
-                    <Field
-                      name="menuAttached"
-                      label={formatMessage(
-                        getTrad('popup.item.form.menuAttached.label', 'MenuAttached')
-                      )}
-                      error={renderError('menuAttached')}
-                      hint={formatMessage(
-                        getTrad(
-                          'popup.item.form.menuAttached.placeholder',
-                          'is menu item attached to menu'
-                        )
-                      )}
-                    >
-                      <Toggle
-                        name="menuAttached"
-                        checked={values.menuAttached}
-                        onChange={(eventOrPath: FormChangeEvent) =>
-                          handleChange(eventOrPath, !values.menuAttached, onChange)
-                        }
-                        value={values.menuAttached}
-                        onLabel={formatMessage(
-                          getTrad('popup.item.form.menuAttached.value.yes', 'yes')
-                        )}
-                        offLabel={formatMessage(
-                          getTrad('popup.item.form.menuAttached.value.no', 'no')
-                        )}
-                        disabled={
-                          !canUpdate ||
-                          (configQuery.data?.cascadeMenuAttached
-                            ? !(current.isMenuAllowedLevel && current.parentAttachedToMenu)
-                            : false)
-                        }
-                        width="100%"
-                      />
-                    </Field>
-                  </Grid.Item>
+                  <NavigationItemTypeField />
+                  <AttachToMenuField current={current} />
                 </Grid.Root>
 
                 <Grid.Root gap={5} paddingTop={1} paddingBottom={1}>
-                  <Grid.Item
-                    alignItems="flex-start"
-                    key="title"
-                    col={values.type === 'INTERNAL' ? 8 : 12}
-                  >
-                    <Field
-                      name="title"
-                      label={formatMessage(getTrad('popup.item.form.title.label', 'Title'))}
-                      error={renderError('title')}
-                      hint={formatMessage(
-                        getTrad('popup.item.form.title.placeholder', 'e.g. Blog')
-                      )}
-                    >
-                      <TextInput
-                        type="string"
-                        disabled={!canUpdate || (values.autoSync && values.type === 'INTERNAL')}
-                        name="title"
-                        onChange={(eventOrPath: FormChangeEvent, value?: any) =>
-                          handleChange(eventOrPath, value, onChange)
-                        }
-                        value={values.title}
-                      />
-                    </Field>
-                  </Grid.Item>
-
-                  {values.type === 'INTERNAL' && (
-                    <Grid.Item alignItems="flex-start" key="autoSync" col={4}>
-                      <Field
-                        name="autoSync"
-                        label={formatMessage(
-                          getTrad('popup.item.form.autoSync.label', 'Read fields from related')
-                        )}
-                        error={renderError('autoSync')}
-                      >
-                        <Toggle
-                          name="autoSync"
-                          checked={values.autoSync}
-                          onChange={(eventOrPath: FormChangeEvent) =>
-                            handleChange(eventOrPath, !values.autoSync, onChange)
-                          }
-                          onLabel="Enabled"
-                          offLabel="Disabled"
-                        />
-                      </Field>
-                    </Grid.Item>
-                  )}
+                  <TitleField />
+                  {values.type === 'INTERNAL' && <ReadFieldsFromRelatedField />}
                 </Grid.Root>
 
                 <Grid.Root gap={5} paddingTop={1} paddingBottom={1}>
-                  <Grid.Item alignItems="flex-start" key="title" col={12}>
-                    <Field
-                      name={pathSourceName}
-                      label={formatMessage(
-                        getTrad(`popup.item.form.${pathSourceName}.label`, 'Path')
-                      )}
-                      error={renderError(
-                        pathSourceName,
-                        `popup.item.form.${pathSourceName}.validation.type`
-                      )}
-                      hint={[
-                        formatMessage(
-                          getTrad(`popup.item.form.${pathSourceName}.placeholder`, 'e.g. Blog')
-                        ),
-                        pathDefault
-                          ? formatMessage(getTrad('popup.item.form.type.external.description'), {
-                              value: pathDefault,
-                            })
-                          : '',
-                      ].join(' ')}
-                    >
-                      <TextInput
-                        disabled={!canUpdate}
-                        name={pathSourceName}
-                        onChange={(eventOrPath: FormChangeEvent, value?: any) =>
-                          handleChange(eventOrPath, value, onChange)
-                        }
-                        // @ts-expect-error
-                        value={values[pathSourceName]}
-                        width="100%"
-                      />
-                    </Field>
-                  </Grid.Item>
+                  <PathField
+                    contentTypeItems={contentTypeItemsQuery.data}
+                    current={current}
+                    isExternal={isExternal}
+                    isSingleSelected={isSingleSelected}
+                  />
                 </Grid.Root>
 
                 {values.type === 'INTERNAL' && (
                   <Grid.Root gap={5} paddingTop={1} paddingBottom={1}>
-                    <Grid.Item
-                      alignItems="flex-start"
-                      col={values.relatedType && !isSingleSelected ? 6 : 12}
-                    >
-                      <Field
-                        name="relatedType"
-                        label={formatMessage(
-                          getTrad('popup.item.form.relatedType.label', 'Related Type')
-                        )}
-                        error={renderError('relatedType')}
-                        hint={
-                          !isLoading && isEmpty(relatedTypeSelectOptions)
-                            ? formatMessage(
-                                getTrad(
-                                  'popup.item.form.relatedType.empty',
-                                  'There are no more content types'
-                                )
-                              )
-                            : undefined
-                        }
-                      >
-                        <ControllableCombobox
-                          name="relatedType"
-                          onClear={() =>
-                            setFormValuesItems({
-                              related: undefined,
-                              title: undefined,
-                              relatedType: undefined,
-                            })
-                          }
-                          onChange={(eventOrPath: FormChangeEvent) =>
-                            setFormValuesItems({
-                              related: undefined,
-                              title: values.autoSync ? '' : values.title,
-                              relatedType: eventOrPath,
-                            })
-                          }
-                          value={values.relatedType}
-                          options={
-                            configQuery.data?.contentTypes.map((contentType) => ({
-                              key: contentType.uid,
-                              value: contentType.uid,
-                              label: contentType.contentTypeName,
-                            })) ?? []
-                          }
-                          disabled={!configQuery.data?.contentTypes.length || !canUpdate}
-                        />
-                      </Field>
-                    </Grid.Item>
-
-                    {values.relatedType && !isSingleSelected && (
-                      <Grid.Item alignItems="flex-start" col={6}>
-                        <Field
-                          name="related"
-                          label={formatMessage(getTrad('popup.item.form.related.label', 'Related'))}
-                          error={renderError('related')}
-                          hint={
-                            !isLoading && thereAreNoMoreContentTypes
-                              ? formatMessage(
-                                  getTrad(
-                                    'popup.item.form.related.empty',
-                                    'There are no more entities'
-                                  ),
-                                  { contentTypeName: values.relatedType }
-                                )
-                              : undefined
-                          }
-                        >
-                          <ControllableCombobox
-                            name="related"
-                            onClear={() => handleChange('related', undefined, onChange)}
-                            onChange={(eventOrPath: FormChangeEvent) =>
-                              handleChange('related', eventOrPath, onChange)
-                            }
-                            value={values.related}
-                            options={relatedSelectOptions}
-                            disabled={isLoading || thereAreNoMoreContentTypes || !canUpdate}
-                          />
-                        </Field>
-                      </Grid.Item>
-                    )}
+                    <RelatedTypeField
+                      contentTypeItems={contentTypeItemsQuery.data}
+                      current={current}
+                      currentRelatedType={values.relatedType}
+                      isSingleSelected={isSingleSelected}
+                      setFormValuesItems={setFormValuesItems}
+                      setIsSingleSelected={setIsSingleSelected}
+                    />
+                    <RelatedEntityField
+                      appendLabelPublicationStatus={appendLabelPublicationStatus}
+                      contentTypeItems={contentTypeItemsQuery.data}
+                      values={values}
+                      isSingleSelected={isSingleSelected}
+                      setFormValuesItems={setFormValuesItems}
+                    />
                   </Grid.Root>
                 )}
 
-                {!isEmpty(configQuery.data?.additionalFields) && (
-                  <>
-                    <Grid.Root gap={5} paddingTop={1} paddingBottom={1}>
-                      <Grid.Item col={12} lg={12}>
-                        <Divider width="100%" />
-                      </Grid.Item>
-                    </Grid.Root>
+                <AdditionalFields />
 
-                    <Grid.Root gap={5} paddingTop={1} paddingBottom={1}>
-                      {configQuery.data?.additionalFields.map(
-                        (additionalField: NavigationItemAdditionalField, index: number) => {
-                          if (additionalField !== 'audience') {
-                            return (
-                              <Grid.Item alignItems="flex-start" key={additionalField.name} col={6}>
-                                <Field
-                                  name={`additionalFields.${additionalField.name}`}
-                                  label={additionalField.label}
-                                  hint={additionalField.description}
-                                  required={additionalField.required}
-                                  error={renderError(`additionalFields.${additionalField.name}`)}
-                                >
-                                  <AdditionalFieldInput
-                                    name={`additionalFields.${additionalField.name}`}
-                                    field={additionalField}
-                                    isLoading={isLoading}
-                                    onChange={onChange}
-                                    onChangeEnhancer={handleChange}
-                                    value={get(values?.additionalFields, additionalField.name)}
-                                    disabled={!canUpdate}
-                                  />
-                                </Field>
-                              </Grid.Item>
-                            );
-                          }
-                        }
-                      )}
-
-                      {configQuery.data?.additionalFields.map(
-                        (additionalField: NavigationItemAdditionalField, index: number) => {
-                          if (additionalField === 'audience') {
-                            return (
-                              <Grid.Item alignItems="flex-start" key="audience" col={12}>
-                                <Field
-                                  name="audience"
-                                  label={formatMessage(getTrad('popup.item.form.audience.label'))}
-                                  error={renderError('audience')}
-                                  hint={
-                                    !isLoading && isEmpty(audienceOptions)
-                                      ? formatMessage(
-                                          getTrad('popup.item.form.title.placeholder', 'e.g. Blog')
-                                        )
-                                      : undefined
-                                  }
-                                >
-                                  <MultiSelect
-                                    name="audience"
-                                    value={values.audience}
-                                    onChange={(eventOrPath: FormChangeEvent) =>
-                                      handleChange('audience', eventOrPath, onChange)
-                                    }
-                                    width="100%"
-                                  >
-                                    {audienceOptions.map(({ value, label }) => (
-                                      <MultiSelectOption key={value} value={value}>
-                                        {label}
-                                      </MultiSelectOption>
-                                    ))}
-                                  </MultiSelect>
-                                </Field>
-                              </Grid.Item>
-                            );
-                          }
-                        }
-                      )}
-                    </Grid.Root>
-
-                    <Grid.Root gap={5} paddingTop={1} paddingBottom={1}>
-                      <Grid.Item col={12} lg={12}>
-                        <Divider width="100%" />
-                      </Grid.Item>
-                    </Grid.Root>
-                  </>
-                )}
-
-                {availableLocaleOptions && availableLocaleOptions.length > 1 && (
-                  <Grid.Root gap={5} paddingTop={1} paddingBottom={1}>
-                    <Grid.Item alignItems="flex-start" key="title" col={12}>
-                      <Divider marginTop={5} marginBottom={5} />
-
-                      <Grid.Root gap={5}>
-                        <Grid.Item alignItems="flex-start" col={6} lg={12}>
-                          <Field
-                            name="i18n.locale"
-                            label={formatMessage(
-                              getTrad('popup.item.form.i18n.locale.label', 'Copy details from')
-                            )}
-                          >
-                            <SingleSelect
-                              name="i18n.locale"
-                              onChange={setItemLocaleCopyValue}
-                              value={itemLocaleCopyValue}
-                              disabled={isLoading || !canUpdate}
-                              placeholder={formatMessage(
-                                getTrad('popup.item.form.i18n.locale.placeholder', 'locale')
-                              )}
-                            >
-                              {availableLocaleOptions.map(({ key, label, value }) => (
-                                <SingleSelectOption key={key} value={value}>
-                                  {label}
-                                </SingleSelectOption>
-                              ))}
-                            </SingleSelect>
-                          </Field>
-                        </Grid.Item>
-
-                        {canUpdate && (
-                          <Grid.Item alignItems="flex-start" col={6} lg={12} paddingTop={6}>
-                            <Box>
-                              <Button
-                                variant="tertiary"
-                                onClick={onCopyFromLocale}
-                                disabled={isLoading || !itemLocaleCopyValue}
-                              >
-                                {formatMessage(getTrad('popup.item.form.i18n.locale.button'))}
-                              </Button>
-                            </Box>
-                          </Grid.Item>
-                        )}
-                      </Grid.Root>
-                    </Grid.Item>
-                  </Grid.Root>
-                )}
+                <CopyFromLocales
+                  availableLocale={availableLocale}
+                  current={current}
+                  currentNavigation={currentNavigation}
+                  setIsLoading={setIsLoading}
+                />
 
                 <Grid.Root gap={5}>
                   <Grid.Item alignItems="flex-start" key="title" col={12}></Grid.Item>
                 </Grid.Root>
-              </>
+              </NavigationItemFormContext.Provider>
             );
           }}
         </Form>
