+++ admin/src/pages/HomePage/components/NavigationItemForm/utils/form.ts
@@ -79,6 +79,9 @@ const navigationItemCommon = ({ additionalFields }: FormSchemaBuilderInput) =>
     ),
   });
 
+export type NavigationInternalItemFormSchema = z.infer<
+  ReturnType<typeof navigationInternalItemFormSchema>
+>;
 const navigationInternalItemFormSchema = ({
   additionalFields,
   isSingleSelected,
