+++ admin/src/pages/HomePage/hooks/useInvalidateQueries.ts
@@ -1,0 +1,12 @@
+import { useEffect } from 'react';
+import { useInvalidateLocale, useInvalidateNavigations } from '.';
+
+export const useInvalidateQueries = () => {
+  const invalidateLocaleQuery = useInvalidateLocale();
+  const invalidateNavigationsQuery = useInvalidateNavigations();
+
+  useEffect(() => {
+    invalidateLocaleQuery();
+    invalidateNavigationsQuery();
+  }, []);
+};
