+++ admin/src/pages/HomePage/index.tsx
@@ -1,274 +1,73 @@
-import {
-  Box,
-  Button,
-  DesignSystemProvider,
-  Flex,
-  SingleSelect,
-  SingleSelectOption,
-  Typography,
-} from '@strapi/design-system';
+import { DesignSystemProvider } from '@strapi/design-system';
 import { Data } from '@strapi/strapi';
-import { Layouts, Page, useNotification, useRBAC } from '@strapi/strapi/admin';
+import { Layouts, Page, useNotification } from '@strapi/strapi/admin';
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
-import { first, isEmpty } from 'lodash';
-import { SyntheticEvent, useCallback, useEffect, useMemo, useState } from 'react';
+import { first } from 'lodash';
+import { SyntheticEvent, useEffect, useState } from 'react';
 import { useIntl } from 'react-intl';
 
 import { usePluginTheme } from '@sensinum/strapi-utils';
-import { ListPlus, Plus } from '@strapi/icons';
-import { NavigationItemSchema, NavigationSchema } from '../../api/validators';
+import { NavigationSchema } from '../../api/validators';
 import { getTrad } from '../../translations';
-import { ToBeFixed } from '../../types';
-import pluginPermissions from '../../utils/permissions';
 import { NavigationHeader } from './components';
 import { ChangeLanguageDialog } from './components/ChangeLanguageDialog';
 import { NavigationContentHeader } from './components/NavigationContentHeader';
 import { NavigationItemFormSchema } from './components/NavigationItemForm';
 import { List } from './components/NavigationItemList';
 import NavigationItemPopUp from './components/NavigationItemPopup';
-import { Search } from './components/Search';
+import { Search } from './components/NavigationContentHeader/Search';
 import {
-  appendViewId,
   useConfig,
-  useCopyNavigationI18n,
-  useI18nCopyNavigationItemsModal,
-  useInvalidateLocale,
-  useInvalidateNavigations,
-  useLocale,
+  useInvalidateQueries,
+  useLocales,
+  useNavigationItemPopup,
   useNavigations,
   usePurgeNavigation,
-  useResetContentTypes,
-  useResetNavigations,
+  useSearch,
+  useSettingsPermissions,
   useUpdateNavigation,
 } from './hooks';
-import {
-  changeCollapseItemDeep,
-  getPendingAction,
-  mapServerNavigationItem,
-  transformItemToViewPayload,
-} from './utils';
+import { getPendingAction, transformItemToViewPayload } from './utils';
+import { ManageNavigationItems } from './components/NavigationContentHeader/ManageNavigationItems';
+import { NavigationEmptyState } from './components/NavigationEmptyState';
+import { appendViewId } from './utils/appendViewId';
 
 const queryClient = new QueryClient();
-
-interface MakeActionInput<T> {
-  trigger(): void;
-  cancel(): void;
-  perform(input: T): void;
-}
-
-const makeAction = <T,>({ cancel, perform, trigger }: MakeActionInput<T>) => {
-  const pointer: { value?: T } = {};
-
-  return {
-    perform: (next?: T) => {
-      const value = next ?? pointer.value;
-
-      if (value) {
-        perform(value);
-      }
-    },
-    trigger: (next: T) => {
-      pointer.value = next;
-
-      trigger();
-    },
-    cancel,
-  };
-};
 
 const Inner = () => {
   const { formatMessage } = useIntl();
 
-  const { toggleNotification } = useNotification();
+  const navigationsQuery = useNavigations();
+  const configQuery = useConfig();
+  const purgeMutation = usePurgeNavigation();
 
-  const localeQuery = useLocale();
+  const { toggleNotification } = useNotification();
 
   const [recentNavigation, setRecentNavigation] = useState<{ documentId?: string; id?: Data.ID }>();
   const [currentNavigation, setCurrentNavigation] = useState<NavigationSchema>();
-
-  const [activeNavigationItem, setActiveNavigationItemState] = useState<
-    Partial<NavigationItemFormSchema> | undefined
-  >();
-
-  const [isItemPopupVisible, setIsItemPopupVisible] = useState(false);
-
   const [structureChanged, setStructureChanged] = useState(false);
 
-  const [currentLocale, setCurrentLocale] = useState<string>();
-  const [isChangeLanguageVisible, setIsChangeLanguageVisible] = useState(false);
-  const changeCurrentLocaleAction = useMemo(
-    () =>
-      makeAction<string>({
-        perform: (next) => {
-          setCurrentLocale(next);
-          setIsChangeLanguageVisible(false);
-          setStructureChanged(false);
-        },
-        trigger: () => setIsChangeLanguageVisible(true),
-        cancel: () => setIsChangeLanguageVisible(false),
-      }),
-    [setCurrentLocale, setIsChangeLanguageVisible]
-  );
-
-  const viewPermissions = useMemo(
-    () => ({
-      access: pluginPermissions.access || pluginPermissions.update,
-      update: pluginPermissions.update,
-    }),
-    []
-  );
+  const { canAccess, canUpdate, isLoadingForPermissions } = useSettingsPermissions();
 
   const {
-    isLoading: isLoadingForPermissions,
-    allowedActions: { canUpdate, canAccess },
-  } = useRBAC(viewPermissions);
+    localeData,
+    currentLocale,
+    isChangeLanguageVisible,
+    changeCurrentLocaleAction,
+    availableLocales,
+  } = useLocales(navigationsQuery.data, setStructureChanged);
 
-  const navigationsQuery = useNavigations();
+  const { searchValue, setSearchValue, isSearchEmpty, filteredList } = useSearch(currentNavigation);
 
-  const configQuery = useConfig();
-
-  const purgeMutation = usePurgeNavigation();
+  const {
+    activeNavigationItem,
+    addNewNavigationItem,
+    editNavigationItem,
+    closeNavigationItemPopup,
+    isItemPopupVisible,
+  } = useNavigationItemPopup(canUpdate);
 
   const pending = getPendingAction([navigationsQuery, { isPending: isLoadingForPermissions }]);
-
-  const copyNavigationI18nMutation = useCopyNavigationI18n();
-
-  const [{ value: searchValue, index: searchIndex }, setSearchValue] = useState({
-    value: '',
-    index: 0,
-  });
-  const isSearchEmpty = isEmpty(searchValue);
-  const normalisedSearchValue = (searchValue || '').toLowerCase();
-
-  const filteredListFactory = (
-    items: Array<NavigationItemSchema>,
-    doUse: (item: NavigationItemSchema) => boolean,
-    activeIndex?: number
-  ): NavigationItemSchema[] => {
-    const filteredItems = items.reduce<Array<NavigationItemSchema>>((acc, item) => {
-      const subItems = !!item.items?.length ? filteredListFactory(item.items ?? [], doUse) : [];
-
-      if (doUse(item)) {
-        return [item, ...subItems, ...acc];
-      } else {
-        return [...subItems, ...acc];
-      }
-    }, []);
-
-    if (activeIndex !== undefined) {
-      const index = activeIndex % filteredItems.length;
-
-      return filteredItems.map((item, currentIndex) => {
-        return index === currentIndex ? { ...item, isSearchActive: true } : item;
-      });
-    }
-
-    return filteredItems;
-  };
-
-  const filteredList = !isSearchEmpty
-    ? filteredListFactory(
-        currentNavigation?.items.map((_) => ({ ..._ })) ?? [],
-        (item) => (item?.title || '').toLowerCase().includes(normalisedSearchValue),
-        normalisedSearchValue ? searchIndex : undefined
-      )
-    : [];
-
-  const changeNavigationItemPopupState = useCallback(
-    (visible: boolean, editedItem = {}) => {
-      setActiveNavigationItemState(editedItem);
-
-      setIsItemPopupVisible(visible);
-    },
-    [setIsItemPopupVisible]
-  );
-
-  const addNewNavigationItem = useCallback(
-    (
-      event: MouseEvent,
-      viewParentId?: number,
-      isMenuAllowedLevel = true,
-      levelPath = '',
-      parentAttachedToMenu = true,
-      structureId = '0',
-      maxOrder = 0
-    ) => {
-      if (canUpdate) {
-        event.preventDefault();
-        event.stopPropagation();
-
-        changeNavigationItemPopupState(true, {
-          viewParentId,
-          isMenuAllowedLevel,
-          levelPath,
-          parentAttachedToMenu,
-          structureId,
-          viewId: undefined,
-          order: maxOrder + 1,
-        });
-      }
-    },
-    [changeNavigationItemPopupState, canUpdate]
-  );
-
-  const availableLocale = useMemo(
-    () =>
-      (localeQuery.data
-        ? [localeQuery.data.defaultLocale, ...localeQuery.data.restLocale]
-        : []
-      ).filter((locale) => locale !== currentLocale),
-    [localeQuery.data, currentLocale]
-  );
-
-  const resetContentTypes = useResetContentTypes();
-  const resetNavigations = useResetNavigations();
-
-  const invalidateLocaleQuery = useInvalidateLocale();
-  const invalidateNavigationsQuery = useInvalidateNavigations();
-
-  const {
-    i18nCopyItemsModal,
-    i18nCopySourceLocale,
-    setI18nCopyModalOpened,
-    setI18nCopySourceLocale,
-  } = useI18nCopyNavigationItemsModal(
-    useCallback(
-      (sourceLocale) => {
-        const source = navigationsQuery.data?.find(
-          ({ locale, documentId }) =>
-            locale === sourceLocale && documentId === currentNavigation?.documentId
-        );
-
-        if (source) {
-          if (source.documentId && currentNavigation?.documentId) {
-            copyNavigationI18nMutation.mutate(
-              {
-                source: source.locale,
-                target: currentNavigation.locale,
-                documentId: source.documentId,
-              },
-              {
-                onSuccess(res) {
-                  copyNavigationI18nMutation.reset();
-                  setCurrentNavigation({
-                    ...res.data,
-                    items: res.data.items.map(appendViewId),
-                  });
-                  resetContentTypes();
-                  resetNavigations();
-                },
-              }
-            );
-          }
-        }
-      },
-      [currentNavigation]
-    )
-  );
-
-  const openI18nCopyModalOpened = useCallback(() => {
-    i18nCopySourceLocale && setI18nCopyModalOpened(true);
-  }, [setI18nCopyModalOpened, i18nCopySourceLocale]);
 
   const updateNavigationMutation = useUpdateNavigation({
     onError: (error: any) => {
@@ -313,75 +112,8 @@ const Inner = () => {
     // the event target element is the root HTML element.
     // This is a workaround to prevent the modal from closing in those cases.
     if ((e.target as any).tagName !== 'HTML') {
-      changeNavigationItemPopupState(false);
+      closeNavigationItemPopup();
     }
-  };
-
-  const handleItemReOrder = ({
-    item,
-    newOrder,
-  }: {
-    item: NavigationItemFormSchema;
-    newOrder: number;
-  }) => {
-    handleSubmitNavigationItem({
-      ...item,
-      order: newOrder,
-    });
-  };
-
-  const handleItemRemove = (item: NavigationItemSchema) => {
-    handleSubmitNavigationItem(
-      mapServerNavigationItem(
-        {
-          ...item,
-          removed: true,
-        },
-        true
-      )
-    );
-  };
-
-  const handleItemRestore = (item: NavigationItemSchema) => {
-    handleSubmitNavigationItem(
-      mapServerNavigationItem(
-        {
-          ...item,
-          removed: false,
-        },
-        true
-      )
-    );
-  };
-
-  const handleItemToggleCollapse = (item: NavigationItemSchema) => {
-    handleSubmitNavigationItem(
-      mapServerNavigationItem(
-        {
-          ...item,
-          collapsed: !item.collapsed,
-          updated: true,
-          isSearchActive: false,
-        },
-        true
-      )
-    );
-  };
-
-  const handleItemEdit = ({
-    item,
-    levelPath = '',
-    parentAttachedToMenu = true,
-  }: {
-    item: NavigationItemFormSchema;
-    levelPath?: string;
-    parentAttachedToMenu?: boolean;
-  }) => {
-    changeNavigationItemPopupState(true, {
-      ...item,
-      levelPath,
-      parentAttachedToMenu,
-    });
   };
 
   const handleSubmitNavigationItem = (payload: NavigationItemFormSchema) => {
@@ -399,57 +131,11 @@ const Inner = () => {
 
       setStructureChanged(true);
 
-      setIsItemPopupVisible(false);
+      closeNavigationItemPopup();
     }
   };
 
   const listItems = isSearchEmpty ? (currentNavigation?.items ?? []) : filteredList;
-
-  const handleExpandAll = useCallback(() => {
-    if (currentNavigation) {
-      setCurrentNavigation({
-        ...currentNavigation,
-        items: currentNavigation.items.map((item) => changeCollapseItemDeep(item, false)),
-      });
-    }
-  }, [setCurrentNavigation, currentNavigation, changeCollapseItemDeep]);
-
-  const handleCollapseAll = useCallback(() => {
-    if (currentNavigation) {
-      setCurrentNavigation({
-        ...currentNavigation,
-        items: currentNavigation.items.map((item) => changeCollapseItemDeep(item, true)),
-      });
-    }
-  }, [setCurrentNavigation, currentNavigation, changeCollapseItemDeep]);
-
-  const endActions = [
-    {
-      onClick: handleExpandAll,
-      type: 'submit',
-      variant: 'tertiary',
-      tradId: 'header.action.expandAll',
-      margin: '8px',
-    },
-    {
-      onClick: handleCollapseAll,
-      type: 'submit',
-      variant: 'tertiary',
-      tradId: 'header.action.collapseAll',
-      margin: '8px',
-    },
-  ] as Array<ToBeFixed>;
-
-  if (canUpdate) {
-    endActions.push({
-      onClick: addNewNavigationItem as ToBeFixed,
-      type: 'submit',
-      variant: 'primary',
-      tradId: 'header.action.newItem',
-      startIcon: <Plus />,
-      margin: '8px',
-    });
-  }
 
   useEffect(() => {
     if (!currentNavigation && navigationsQuery.data?.[0]) {
@@ -485,19 +171,9 @@ const Inner = () => {
     }
   }, [currentNavigation, currentLocale, navigationsQuery.data]);
 
-  useEffect(() => {
-    if (!currentLocale && localeQuery.data?.defaultLocale) {
-      setCurrentLocale(localeQuery.data?.defaultLocale);
-      setStructureChanged(false);
-    }
-  }, [navigationsQuery.data]);
+  useInvalidateQueries();
 
-  useEffect(() => {
-    invalidateLocaleQuery();
-    invalidateNavigationsQuery();
-  }, []);
-
-  if (!navigationsQuery.data || !localeQuery.data || !!pending) {
+  if (!navigationsQuery.data || !localeData || !!pending) {
     return <Page.Loading />;
   }
 
@@ -514,7 +190,7 @@ const Inner = () => {
             structureChanged ? changeCurrentLocaleAction.trigger : changeCurrentLocaleAction.perform
           }
           handleSave={submit}
-          locale={localeQuery.data}
+          locale={localeData}
           structureHasChanged={structureChanged}
           isSaving={updateNavigationMutation.isPending}
           permissions={{ canUpdate }}
@@ -524,73 +200,30 @@ const Inner = () => {
         <Layouts.Content>
           <NavigationContentHeader
             startActions={<Search value={searchValue} setValue={setSearchValue} />}
-            endActions={endActions.map(({ tradId, margin, ...item }, i) => (
-              <Box marginLeft={margin} key={i}>
-                <Button {...item}> {formatMessage(getTrad(tradId))} </Button>
-              </Box>
-            ))}
+            endActions={
+              <ManageNavigationItems
+                currentNavigation={currentNavigation}
+                setCurrentNavigation={setCurrentNavigation}
+                canUpdate={canUpdate}
+                addNewNavigationItem={addNewNavigationItem}
+              />
+            }
           />
           {!currentNavigation?.items.length ? (
-            <Flex direction="column" minHeight="400px" justifyContent="center">
-              <Box padding={4}>
-                <Typography variant="beta" textColor="neutral600">
-                  {formatMessage(getTrad('empty.description'))}
-                </Typography>
-              </Box>
-              {canUpdate && (
-                <Button
-                  variant="secondary"
-                  startIcon={<ListPlus />}
-                  label={formatMessage(getTrad('empty.cta'))}
-                  onClick={addNewNavigationItem}
-                >
-                  {formatMessage(getTrad('empty.cta'))}
-                </Button>
-              )}
-              {canUpdate && availableLocale.length ? (
-                <Flex direction="column" justifyContent="center">
-                  <Box paddingTop={3} paddingBottom={3}>
-                    <Typography variant="beta" textColor="neutral600">
-                      {formatMessage(getTrad('view.i18n.fill.cta.header'))}
-                    </Typography>
-                  </Box>
-                  <Flex direction="row" justifyContent="center" alignItems="center">
-                    <Box paddingLeft={1} paddingRight={1}>
-                      <SingleSelect
-                        onChange={setI18nCopySourceLocale}
-                        value={i18nCopySourceLocale}
-                        size="S"
-                      >
-                        {availableLocale.map((locale) => (
-                          <SingleSelectOption key={locale} value={locale}>
-                            {formatMessage(getTrad('view.i18n.fill.option'), { locale })}
-                          </SingleSelectOption>
-                        ))}
-                      </SingleSelect>
-                    </Box>
-                    <Box paddingLeft={1} paddingRight={1}>
-                      <Button
-                        variant="tertiary"
-                        onClick={openI18nCopyModalOpened}
-                        disabled={!i18nCopySourceLocale}
-                        size="S"
-                      >
-                        {formatMessage(getTrad('view.i18n.fill.cta.button'))}
-                      </Button>
-                    </Box>
-                  </Flex>
-                </Flex>
-              ) : null}
-            </Flex>
+            <NavigationEmptyState
+              canUpdate={canUpdate}
+              addNewNavigationItem={addNewNavigationItem}
+              availableLocale={availableLocales}
+              availableNavigations={navigationsQuery.data}
+              currentNavigation={currentNavigation}
+              setCurrentNavigation={setCurrentNavigation}
+            />
           ) : (
             <List
               items={listItems}
               onItemLevelAdd={addNewNavigationItem}
-              onItemRemove={handleItemRemove}
-              onItemEdit={handleItemEdit}
-              onItemRestore={handleItemRestore}
-              onItemReOrder={handleItemReOrder}
-              onItemToggleCollapse={handleItemToggleCollapse}
+              onItemEdit={editNavigationItem}
+              onItemSubmit={handleSubmitNavigationItem}
               displayFlat={!isSearchEmpty}
               isParentAttachedToMenu
               permissions={{ canUpdate, canAccess }}
@@ -601,7 +234,7 @@ const Inner = () => {
 
           {isItemPopupVisible && currentLocale && currentNavigation && (
             <NavigationItemPopUp
-              availableLocale={availableLocale}
+              availableLocale={availableLocales}
               currentItem={activeNavigationItem}
               onSubmit={handleSubmitNavigationItem}
               onClose={onPopUpClose}
@@ -613,14 +246,12 @@ const Inner = () => {
             />
           )}
 
-          {isChangeLanguageVisible ? (
+          {isChangeLanguageVisible && (
             <ChangeLanguageDialog
               onCancel={() => changeCurrentLocaleAction.cancel()}
               onConfirm={() => changeCurrentLocaleAction.perform()}
             />
-          ) : null}
-
-          {canUpdate && i18nCopyItemsModal}
+          )}
         </Layouts.Content>
       </Page.Main>
     </Layouts.Root>
